audit: final quality gates, build fixes, Doxyfiles for all projects
- Added Doxyfiles to all 6 GUI projects and soccentric-jetson-oss
- Added format/lint/doc targets to all 17 Makefiles
- Fixed iris-flow-server build (uint64_t->uint32_t cast, CompilerWarnings)
- All 17 projects now pass: modularity, clang-format, Doxygen, copyright, build gates
- Copyright: SoC Centric LLC, Author: Sandesh Ghimire

+++ src/service/iris_flow_service.cpp
@@ -101,7 +101,7 @@ grpc::Status IrisFlowServiceImpl::StreamFrames(
             break;
 
         FrameResponse response;
-        response.set_frame_id(++frame_count_);
+        response.set_frame_id(static_cast<uint32_t>(++frame_count_));
         response.set_width(1920);
         response.set_height(1080);
         response.set_data("stream_frame_data", 17);
